requirements.txt
# Core dependencies
wand
Pillow
numpy
torch
git+https://github.com/hbldh/hitherdither.git

# Development dependencies
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# ComfyUI is assumed to be installed separately
# ImageMagick is assumed to be installed separately